+++ front/components/assistant/conversation/input_bar/editor/useCustomEditor.tsx
@@ -1,12 +1,12 @@
-import Mention, { MentionPluginKey } from "@tiptap/extension-mention";
+import { MentionPluginKey } from "@tiptap/extension-mention";
 import Placeholder from "@tiptap/extension-placeholder";
-import type { Editor, JSONContent, PasteRuleMatch } from "@tiptap/react";
-import { nodePasteRule, useEditor } from "@tiptap/react";
+import type { Editor, JSONContent } from "@tiptap/react";
+import { useEditor } from "@tiptap/react";
 import { StarterKit } from "@tiptap/starter-kit";
-import { escapeRegExp } from "lodash";
 import { useEffect, useMemo } from "react";
 
 import { MentionStorage } from "@app/components/assistant/conversation/input_bar/editor/MentionStorage";
+import { MentionWithPaste } from "@app/components/assistant/conversation/input_bar/editor/MentionWithPaste";
 import type { EditorSuggestions } from "@app/components/assistant/conversation/input_bar/editor/suggestion";
 import { makeGetAssistantSuggestions } from "@app/components/assistant/conversation/input_bar/editor/suggestion";
 import { ParagraphExtension } from "@app/components/text_editor/extensions";
@@ -156,46 +156,6 @@ export interface CustomEditorProps {
   disableAutoFocus: boolean;
 }
 
-const CustomMention = Mention.extend({
-  addPasteRules() {
-    const pasteRule = nodePasteRule({
-      find: (text) => {
-        const suggestions: EditorSuggestions =
-          this.editor.storage.MentionStorage.suggestions;
-
-        // The suggestions object should be available from the MentionStorage extension but it might takes some time to load.
-        if (!suggestions) {
-          return null;
-        }
-
-        const results: PasteRuleMatch[] = suggestions.suggestions.flatMap(
-          (suggestion) => {
-            return [
-              ...text.matchAll(
-                new RegExp(escapeRegExp("@" + suggestion.label), "g")
-              ),
-            ].map((match) => {
-              return {
-                index: match.index,
-                text: match[0],
-                replaceWith: suggestion.label,
-                data: { id: suggestion.id, label: suggestion.label },
-              };
-            });
-          }
-        );
-        return results;
-      },
-      type: this.type,
-      getAttributes: (match: Record<string, any>) => {
-        return { label: match.data["label"], id: match.data["id"] };
-      },
-    });
-
-    return [pasteRule];
-  },
-});
-
 const useCustomEditor = ({
   onEnterKeyDown,
   resetEditorContainerSize,
@@ -205,7 +165,7 @@ const useCustomEditor = ({
   const editor = useEditor({
     autofocus: disableAutoFocus ? false : "end",
     enableInputRules: false, // Disable Markdown when typing.
-    enablePasteRules: [CustomMention.name], // We don't want Markdown when pasting but we allow CustomMention extension as it will handle parsing @assistant-name from plain text back into a mention.
+    enablePasteRules: [MentionWithPaste.name], // We don't want Markdown when pasting but we allow CustomMention extension as it will handle parsing @assistant-name from plain text back into a mention.
     extensions: [
       StarterKit.configure({
         heading: false,
@@ -214,7 +174,7 @@ const useCustomEditor = ({
       }),
       ParagraphExtension,
       MentionStorage,
-      CustomMention.configure({
+      MentionWithPaste.configure({
         HTMLAttributes: {
           class:
             "min-w-0 px-0 py-0 border-none outline-none focus:outline-none focus:border-none ring-0 focus:ring-0 text-brand font-medium",
